requirements.txt
opencv-python
pyzbar
pandas
fpdf
python-barcode
Pillow